+++ docs/docs/zh/CHANGELOG.md
@@ -5,7 +5,7 @@
 ## 3.4.12 (2026-06-05)
 
 - BREAKECHANGE
-  - 针对 3.4.11, mdit2xxx (list 除外) 默认将尽可能地使用非 fine 模式
+  - mdit 和 title 转 table 时，将默认使用非 fine 模式 (以前是 fine 模式)
 - feat
   - 新增别名: `h2xxx` 和 `heading2xxx` 表示 `title2xxx`
   - `title2listdata` 和 `mdit2listdata` 将分别支持 fine 模式和 非 fine 模式